Move test out of the library and into tests folder
